+++ src/main/resources/com/axis/system/jenkins/plugins/downstream/yabv/BuildFlowAction/jobMain.groovy
@@ -1,4 +1,6 @@
 package com.axis.system.jenkins.plugins.downstream.yabv.BuildFlowAction
 
-h2('Build Flow')
-include(my, 'buildFlowJsCss.groovy')
+if (my.hasUpstreamOrDownstreamBuilds()) {
+  h2('Build Flow')
+  include(my, 'buildFlowJsCss.groovy')
+}