+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## v1.9:
-* **Feature**: Toggle showing build execution times.
+* **Feature**: Switch to show build execution times.
+* **Feature**: Switch to show dots indicating the build results of the 5 previous builds in each cell.
 * **Feature**: Fullscreen Build Flow. Append "/yabv" to project/build page url.
 
 ## v1.8.1: